+++ components/Directory.jsx
@@ -1,15 +1,11 @@
 import { useState } from 'react';
-import Avatar from '@mui/material/Avatar';
 import Box from '@mui/material/Box';
 import Grid from '@mui/material/Grid';
-import Typography from '@mui/material/Typography';
 import Tab from '@mui/material/Tab';
 import Tabs from '@mui/material/Tabs';
 import { motion } from 'framer-motion'
+import PersonCard from '../components/PersonCard'
 import { slugify } from '../src/utils';
-import defaultProfile from '../public/default-profile.png'
-import Chip from '@mui/material/Chip';
-import { blue, green, red } from '@mui/material/colors'
 
 export default function Directory({ items }) {
   const [value, setValue] = useState(0);
@@ -32,7 +28,7 @@ export default function Directory({ items }) {
                     whileInView={{ opacity: 1, y: 0 }}
                     transition={{ delay: 0.05 * index }}
                   >
-                    <Person {...person} />
+                    <PersonCard {...person} />
                   </motion.div>
                 </Grid>
               );
@@ -79,48 +75,3 @@ const TabPanel = ({ children, value, index, ...other }) => (
   </div>
 );
 
-function getCategoryColor(categoryName) {
-  switch (categoryName) {
-    case 'Psych': return blue[300];
-    case 'Nexus': return red[400];
-    case 'Telemedicine': return green[400];
-    default: return;
-  }
-}
-
-const Person = ({ name, position, image, isTeamLead, team }) => {
-  const imageUrl = image && Array.isArray(image) ? image[0] : image ? image : defaultProfile.src;
-  const srcset = image && Array.isArray(image) && image.length > 0 ? image.join(', ') : null;
-  const label = `${team} Team${isTeamLead ? ' Lead' : ''}`;
-
-  return <>
-    <Grid container spacing={{ xs: 2, md: 1 }} alignItems={'center'}>
-      <Grid item md={12}>
-        <Avatar srcSet={srcset} src={imageUrl} alt={`${name} profile photo`} sx={{ width: {xs: 72, md: 150}, height: {xs: 72, md: 150}, marginBottom: 1, mx: 'auto' }} />        
-      </Grid>
-      <Grid item xs md={12} sx={{ textAlign: { xs: 'left', md: 'center' } }}>
-        <Box sx={[ team && position ? { display: ['flex', 'flex', 'initial']} : null ]}>
-          <Typography variant='h6' component='p' gutterBottom sx={{ lineHeight: 1 }}>{name}</Typography>
-          {position ? (
-            <Typography variant='body2' sx={[ team ? { ml: [1, 1, 0] } : null ]}>{position}</Typography>
-          ) : null}
-        </Box>
-        {team ? (
-          <Chip
-            label={label}
-            size='small'
-            variant={isTeamLead ? 'contained' : 'outlined'}
-            sx={[
-              { color: isTeamLead ? '#fff' : getCategoryColor(team) },
-              isTeamLead
-                ? { backgroundColor: getCategoryColor(team) }
-                : { borderColor: getCategoryColor(team) },
-            ]}
-          />
-        ) : null}
-      </Grid>
-    </Grid>
-    
-    
-  </>
-};
